+++ CHANGELOG.md
@@ -11,13 +11,21 @@ All user visible changes to this project will be documented in this file. This p
 
 [Diff](/../../compare/v0.6.6...v0.7.0) | [Milestone](/../../milestone/51)
 
+### Added
+
+- UI:
+    - Pages scrolling by pressing "PageUp" and "PageDown" keys. ([#1469], [#1228])
+
+
 ### Fixed
 
 - UI:
     - Media panel:
         - Invalid tooltip positions for buttons in dock. ([#1473], [#1436])
 
+[#1228]: /../../issues/1228
 [#1436]: /../../issues/1436
+[#1469]: /../../pull/1469
 [#1473]: /../../pull/1473